+++ libs/cardano-ledger-api/src/Cardano/Ledger/Api/Tx.hs
@@ -1,3 +1,8 @@
+{-# LANGUAGE DataKinds #-}
+{-# LANGUAGE ScopedTypeVariables #-}
+{-# LANGUAGE TypeApplications #-}
+{-# LANGUAGE TypeOperators #-}
+
 -- | Transaction building and inspecting relies heavily on lenses (`microlens`). Therefore, some
 -- familiarity with those is necessary. However, you can probably go a long way by simply
 -- looking at the examples and try to go from there.
@@ -43,6 +48,7 @@ module Cardano.Ledger.Api.Tx (
   sizeTxF,
   getMinFeeTx,
   setMinFeeTx,
+  estimateMinFeeTx,
 
   -- * Alonzo onwards
   AlonzoEraTx,
@@ -58,6 +64,8 @@ module Cardano.Ledger.Api.Tx (
 )
 where
 
+import qualified Cardano.Chain.Common as Byron
+import Cardano.Crypto.DSIGN.Class (sizeSigDSIGN, sizeVerKeyDSIGN)
 import Cardano.Ledger.Alonzo.Tx (AlonzoEraTx (..), IsValid (..))
 import Cardano.Ledger.Api.Era ()
 import Cardano.Ledger.Api.Scripts.ExUnits (
@@ -71,4 +79,93 @@ import Cardano.Ledger.Api.Tx.AuxData
 import Cardano.Ledger.Api.Tx.Body
 import Cardano.Ledger.Api.Tx.Cert
 import Cardano.Ledger.Api.Tx.Wits
-import Cardano.Ledger.Core (EraTx (..), setMinFeeTx)
+import Cardano.Ledger.BaseTypes (ProtVer (..))
+import Cardano.Ledger.Binary (byronProtVer, decodeFull', serialize')
+import Cardano.Ledger.Coin (Coin)
+import Cardano.Ledger.Core (EraCrypto, EraTx (..), PParams, ppProtocolVersionL, setMinFeeTx)
+import Cardano.Ledger.Crypto
+import Cardano.Ledger.Keys (BootstrapWitness (..), ChainCode (..), VKey)
+import Data.Bits (shiftR)
+import qualified Data.ByteString as BS
+import Data.Proxy
+import qualified Data.Set as Set
+import Lens.Micro
+
+-- | Estimate a minimum transaction fee for a transaction that does not yet have all of
+-- the `VKey` witnesses. This calculation is not very accurate in estimating Byron
+-- witnesses, but it should work for the most part.
+--
+-- @since 1.8.0
+estimateMinFeeTx ::
+  forall era.
+  EraTx era =>
+  -- | The current protocol parameters.
+  PParams era ->
+  -- | The transaction.
+  Tx era ->
+  -- | The number of key witnesses still to be added to the transaction.
+  Int ->
+  -- | The number of Byron key witnesses still to be added to the transaction.
+  Int ->
+  -- | The required minimum fee.
+  Coin
+estimateMinFeeTx pp tx numKeyWits numByronKeyWits = setMinFeeTx pp tx' ^. bodyTxL . feeTxBodyL
+  where
+    tx' = addDummyWitsTx pp tx numKeyWits $ replicate numByronKeyWits dummyByronAttributes
+    -- We assume testnet network magic here to avoid having to thread the actual network
+    -- ID into this function merely to calculate the fees of byron witnesses more
+    -- accurately. This will over-estimate min fees for byron witnesses in mainnet
+    -- transaction by 7 bytes per witness.
+    dummyByronAttributes =
+      Byron.AddrAttributes
+        { Byron.aaVKDerivationPath = Nothing
+        , Byron.aaNetworkMagic = Byron.NetworkTestnet maxBound
+        }
+
+addDummyWitsTx ::
+  forall era.
+  EraTx era =>
+  -- | The current protocol parameters.
+  PParams era ->
+  -- | The transaction.
+  Tx era ->
+  -- | The number of key witnesses still to be added to the transaction.
+  Int ->
+  -- | List of attributes from TxOuts with Byron addresses that are being spent
+  [Byron.AddrAttributes] ->
+  -- | The required minimum fee.
+  Tx era
+addDummyWitsTx pp tx numKeyWits byronAttrs =
+  tx
+    & (witsTxL . addrTxWitsL <>~ dummyKeyWits)
+    & (witsTxL . bootAddrTxWitsL <>~ dummyByronKeyWits)
+  where
+    dsign :: Proxy (DSIGN (EraCrypto era))
+    dsign = Proxy
+    version = pvMajor (pp ^. ppProtocolVersionL)
+    -- We need to make sure that dummies are unique, since they'll be placed into a Set
+    mkDummy name n =
+      either
+        (\err -> error ("Corrupt Dummy " ++ name ++ ": " ++ show err))
+        id
+        . decodeFull' version
+        . serialize' version
+        . fst
+        . BS.unfoldrN n (\b -> Just (fromIntegral b, b `shiftR` 8))
+    vKeySize = fromIntegral $ sizeVerKeyDSIGN dsign
+    dummyKeys = map (mkDummy "VKey" vKeySize) [0 :: Int ..]
+
+    sigSize = fromIntegral $ sizeSigDSIGN dsign
+    dummySig = mkDummy "Signature" sigSize (0 :: Int)
+    dummyKeyWits =
+      Set.fromList [WitVKey key dummySig | key <- take numKeyWits dummyKeys]
+
+    -- ChainCode is always 32 bytes long.
+    chainCode = ChainCode $ BS.replicate 32 0
+
+    mkDummyByronKeyWit ::
+      VKey 'Witness (EraCrypto era) -> Byron.AddrAttributes -> BootstrapWitness (EraCrypto era)
+    mkDummyByronKeyWit key =
+      BootstrapWitness key dummySig chainCode . serialize' byronProtVer . Byron.mkAttributes
+    dummyByronKeyWits =
+      Set.fromList $ zipWith mkDummyByronKeyWit dummyKeys byronAttrs