+++ libs/cardano-ledger-api/test/Tests.hs
@@ -5,11 +5,11 @@ module Main where
 import Cardano.Ledger.Conway (Conway)
 import qualified Test.Cardano.Ledger.Api.State.Imp.QuerySpec as ImpQuery (spec)
 import qualified Test.Cardano.Ledger.Api.State.QuerySpec as StateQuery (spec)
+import qualified Test.Cardano.Ledger.Api.Tx as Tx (spec)
 import qualified Test.Cardano.Ledger.Api.Tx.Body as TxBody (spec)
 import qualified Test.Cardano.Ledger.Api.Tx.Out as TxOut (spec)
-import Test.Cardano.Ledger.Conway.ImpTest (withImpState)
-
 import Test.Cardano.Ledger.Common
+import Test.Cardano.Ledger.Conway.ImpTest (withImpState)
 
 -- ====================================================================================
 
@@ -17,6 +17,7 @@ apiSpec :: Spec
 apiSpec =
   describe "cardano-ledger-api" $ do
     describe "Tx" $ do
+      Tx.spec
       TxOut.spec
       TxBody.spec
     describe "State" $ do