+++ libs/cardano-ledger-binary/CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## 1.3.0.0
 
+* Add instance `EncCBOR`/`DecCBOR` instance for `SignedDSIGN` and add `encodedSignedDSIGNSizeExpr`
 * Remove unused and badly named `decodeMapNoDuplicates`
 * Fix `decodeVMap` to no longer allow duplicates
 * Add `decodeMapLikeEnforceNoDuplicates` and `decodeListLikeWithCount`